+++ src/System/Taffybar/Information/Battery.hs
@@ -199,7 +199,7 @@ getDisplayBatteryInfo = do
 -- | Default set of UPower properties that trigger display-battery refreshes.
 defaultMonitorDisplayBatteryProperties :: [String]
 defaultMonitorDisplayBatteryProperties =
-  ["IconName", "State", "Percentage", "EnergyRate"]
+  ["IconName", "State", "Percentage", "EnergyRate", "UpdateTime"]
 
 -- | Start the monitoring of the display battery, and setup the associated
 -- channel and mvar for the current state.